Added newPseudonym end point
