+++ src/routes/v1/user.route.js
@@ -6,6 +6,6 @@ const validate = require('../../middlewares/validate');
 const router = express.Router();
 
 router.post('/', validate(userValidation.createUser), userController.createUser);
-router.route('/newToken').get(userController.newToken);
+router.route('/newPseudonym').get(userController.newPseudonym);
 
 module.exports = router;
